Changed the lib compiling process & introduced possible solution for #52

+++ src/lib/stdioeds.c
@@ -3,7 +3,7 @@
 #include <stdint.h>
 #include <inttypes.h>
 
-extern void writeInteger_i_0 (int n){
+extern void writeInteger_i_0 (int16_t n){
 	printf("%hd", n);
 }
